+++ Example/CustomBulletins/DatePickerBulletinItem.swift
@@ -10,17 +10,15 @@ import BLTNBoard
  * A bulletin item that demonstrates how to integrate a date picker inside a bulletin item.
  */
 
-class DatePickerBLTNItem: BLTNPageItem {
-
-    lazy var datePicker = UIDatePicker()
+@objc public class DatePickerBLTNItem: BLTNPageItem {
+    public lazy var datePicker = UIDatePicker()
 
     /**
      * Display the date picker under the description label.
      */
-
-    override func makeViewsUnderDescription(with interfaceBuilder: BLTNInterfaceBuilder) -> [UIView]? {
+    
+    override public func makeViewsUnderDescription(with interfaceBuilder: BLTNInterfaceBuilder) -> [UIView]? {
         datePicker.datePickerMode = .date
         return [datePicker]
     }
-
 }