+++ Sources/BLTNItemManager.swift
@@ -181,11 +181,8 @@ extension BLTNItemManager {
         bulletinController.transitioningDelegate = bulletinController
         bulletinController.loadBackgroundView()
         bulletinController.setNeedsStatusBarAppearanceUpdate()
-
-        if #available(iOS 11.0, *) {
-            bulletinController.setNeedsUpdateOfHomeIndicatorAutoHidden()
-        }
-
+        bulletinController.setNeedsUpdateOfHomeIndicatorAutoHidden()
+        
         isPrepared = true
         isPreparing = true
         shouldDisplayActivityIndicator = rootItem.shouldStartWithActivityIndicator
@@ -303,7 +300,8 @@ extension BLTNItemManager {
      * - parameter item: The item to display.
      */
 
-    @objc public func push(item: BLTNItem) {
+    @objc(pushItem:)
+    public func push(item: BLTNItem) {
 
         assertIsPrepared()
         assertIsMainThread()
@@ -352,7 +350,8 @@ extension BLTNItemManager {
      * - parameter orDismiss: If true, dismiss bullein if not found. Otherwise popToRootItem()
      */
     
-    @objc public func popTo(item: BLTNItem, orDismiss: Bool) {
+    @objc(popToItem:orDismiss:)
+    public func popTo(item: BLTNItem, orDismiss: Bool) {
         
         assertIsPrepared()
         assertIsMainThread()