+++ Sources/InterfaceBuilder/BLTNInterfaceBuilder.swift
@@ -12,23 +12,20 @@ import UIKit
 
 @objc open class BLTNInterfaceBuilder: NSObject {
 
-    /// The item for which the interface builder was created.
-    @objc public weak var item: BLTNItem?
-
     /// The appearance to use to generate the items.
     @objc public let appearance: BLTNItemAppearance
 
     /// Creates a new interface builder.
-    @objc public required init(appearance: BLTNItemAppearance, item: BLTNItem) {
+    @objc public required init(appearance: BLTNItemAppearance) {
         self.appearance = appearance
-        self.item = item
     }
 
     /**
      * Creates a standard title label.
      */
 
-    @objc open func makeTitleLabel() -> BLTNTitleLabelContainer {
+    @objc(makeTitleLabelNextToCloseButton:)
+    open func makeTitleLabel(isNextToCloseButton: Bool) -> BLTNTitleLabelContainer {
 
         let titleLabel = UILabel()
         titleLabel.textAlignment = .center
@@ -39,12 +36,9 @@ import UIKit
         titleLabel.lineBreakMode = .byWordWrapping
 
         titleLabel.font = appearance.makeTitleFont()
-
-        let needsCloseButton = item?.isDismissable == true && item?.requiresCloseButton == true
-        let inset: CGFloat = needsCloseButton ? 12 + 30 : 0
+        let inset: CGFloat = isNextToCloseButton ? 12 + 30 : 0
 
         return BLTNTitleLabelContainer(label: titleLabel, horizontalInset: inset)
-
     }
 
     /**
@@ -98,7 +92,11 @@ import UIKit
     @objc open func makeActionButton(title: String) -> BLTNHighlightButtonWrapper {
 
         let actionButton = HighlightButton()
-        actionButton.cornerRadius = appearance.actionButtonCornerRadius
+        actionButton.layer.cornerRadius = appearance.actionButtonCornerRadius
+        
+        if #available(iOS 13, *) {
+            actionButton.layer.cornerCurve = .continuous
+        }
         
         if let actionButtonImage = appearance.actionButtonImage {
             actionButton.setBackgroundImage(actionButtonImage, for: .normal)
@@ -141,8 +139,13 @@ import UIKit
 
     @objc open func makeAlternativeButton(title: String) -> UIButton {
 
-        let alternativeButton = RoundedButton()
-        alternativeButton.cornerRadius = appearance.alternativeButtonCornerRadius
+        let alternativeButton = UIButton()
+        alternativeButton.layer.cornerRadius = appearance.alternativeButtonCornerRadius
+        
+        if #available(iOS 13, *) {
+            alternativeButton.layer.cornerCurve = .continuous
+        }
+        
         alternativeButton.setTitle(title, for: .normal)
         alternativeButton.setTitleColor(appearance.alternativeButtonTitleColor, for: .normal)
         alternativeButton.titleLabel?.font = appearance.makeAlternativeButtonFont()