+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # _BulletinBoard_ Changelog
 ## Unreleased
+### Changes
+- Require iOS 11.0
+- Support for Swift Package Manager
+
+### Fixes
+- Fix the background view origin when presenting
+[#183](https://github.com/alexaubry/BulletinBoard/pull/183)
 
 ## 🔖 v4.1.2
 ### Fixes